allinall mapstructure, fanzy includes and modules for later
