+++ network_project/client/drivers/connect.h
@@ -2,8 +2,8 @@
 #define CONNECT_H_
 
 #include <stdint.h>
-#include "utils/cconfig.h"
-#include "../client.h"
+#include "cconfig.h"
+#include "client.h"
 
 int8_t connect_driver(client_t *client, int8_t choice);
