+++ network_project/client/drivers/cstring.c
