+++ network_project/client/drivers/cstring.h
