+++ network_project/client/drivers/fetcher.c
@@ -1,21 +1,17 @@
 /*------------------------------------------------------------------------------------------------------------------------
-                                                                                                                    DEVICE
+                                                                                                                   FETCHER
 --------------------------------------------------------------------------------------------------------------------------
 info info info info info info
 ------------------------------------------------------------------------------------------------------------------------*/
-#include <string.h>
-#include <stdio.h>
-#include "utils/cstring.h"
-#include "utils/scanner.h"
-#include "device.h"
+#include "cstring.h"
+#include "scanner.h"
+#include "fetcher.h"
 
-int8_t device_driver(client_t *client, int8_t choice) {
-  if (choice == DRED)
-    client->rqst[0] = RADR;
-  else if (choice == DBLU)
-    client->rqst[0] = RADB;
+int8_t fetch_driver(client_t *client, int8_t choice) {
+  if (choice == TEMP)
+    client->rqst[0] = RDTP;
   else
-    client->rqst[0] = RADG;
+    client->rqst[0] = RDMG;
   client->size_rqst = 2;
   return DONE;
 }
