+++ network_project/client/drivers/fetcher.h
@@ -2,8 +2,8 @@
 #define FETCHER_H_
 
 #include <stdint.h>
-#include "utils/cconfig.h"
-#include "../client.h"
+#include "cconfig.h"
+#include "client.h"
 
 int8_t fetch_driver(client_t *client, int8_t choice);
