+++ network_project/client/drivers/scanner.c
