+++ network_project/client/makefile
@@ -1,5 +1,5 @@
 make:
-	gcc -o clientTCP clientTCP.c socket.c request/*c request/reqdrivers/*c request/reqdrivers/utils/*c -Wall -Wextra -g
+	gcc -o clientTCP *c drivers/*c -Wall -Wextra -g
 
 clean:
 	rm -f $(obj) clientTCP
